backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
aiofiles
openai
chromadb
PyPDF2
feedparser
python-multipart